Restructure lib folder to allow for more modules.
