+++ lib/specifications/projects/project.js
@@ -3,12 +3,12 @@
 
 var fs = require('q-io/fs');
 
-var getProjectGit = require('./repository-types/git').getProject;
-var getProjectDataGit = require('./repository-types/git').getProjectData;
-var getFileContentsGit = require('./repository-types/git').getFileContents;
+var getProjectGit = require('../repository-types/git').getProject;
+var getProjectDataGit = require('../repository-types/git').getProjectData;
+var getFileContentsGit = require('../repository-types/git').getFileContents;
 
-var appConfig = require('../configuration').get();
-var projectConfig = require('../configuration/project-config');
+var appConfig = require('../../configuration/app-config').get();
+var projectConfig = require('../../configuration/project-config');
 
 
 /**
